Booth Placement Algorithm V2
Fixes some issues with the booth placement algorithm. Now confirmed to work with a 3x3 matrix of spaces, as well as with non-available spaces. Not currently tested with any matrices larger than 3x3 or non-square matrices, although it should work correctly. A small change was also made to the algorithm for the buckets of size 3, 2 and 1. Javadocs for the BoothPlacer file are now also included.